User registration (#2)
Added user registration support. 

When users are created by sending POST requests to the /users/ endpoint, an email with an activation code is sent to the user's email address. The user then visits /activate/?token=[token] to activate their account.

Templates for environment variable settings are included in .env. These are as for python's `smtplib` via Django's `send_email`.