Storage (#31)
Galv now offers storage to Labs. This storage can be on the server's local filesystem, or it can be an AWS S3 bucket. The quota for the storage is set via environment variable. Galv storage can be disabled by a Lab.

Additional storages can be configured on a Lab. These are S3 buckets using the Lab's credentials. Quotas can be set on these storages and adjusted later. 

Storages, including Galv Storage, can be enabled or disabled for the Lab, and their priorities can be adjusted. Files will be saved to the highest priority storage with sufficient space remaining in its quota when they are first saved. Files cannot be moved between storages. If a storage is disabled, files will still be available to retrieve. If a storage is deleted, files will be unavailable.

Storages are used for the PNG overview of file data, and for the parquet files that hold file data.